Insertion of Alliance/Poi  Player/Base

+++ src/main/java/org/cnc/cncbot/dto/publicplayerinfo/PublicPlayerInfoResponse.java
@@ -35,7 +35,7 @@ private Integer i;
 
 private String n;
 
-private Integer p;
+private Long p;
 
 private Integer r;
