+++ src/main/java/org/cnc/cncbot/stats/entities/StatsLog.java
@@ -46,6 +46,6 @@ public class StatsLog {
 	private Integer world;
 	
 	@Column(name = "alliance")
-	private Long alliance;
+	private Integer alliance;
 	
 }
